Converted coordinate variables to float to fix gem group to gem grid coordinate discrepancy.

+++ app/src/main/java/com/jcrawleydev/gemsdrop/control/GemControls.java
@@ -142,7 +142,7 @@ public class GemControls {
 
 
     private boolean doesColumnHeightMeetMiddleGem(int colIndex){
-        int gemGroupMidY = gemGroup.getBottomPosition() + 1;
+        float gemGroupMidY = gemGroup.getBottomPosition() + 1;
         return getColumnTopPosition(colIndex) >= gemGroupMidY;
     }
